UPSTREAM: ASoC: cs42l42: Remove redundant writes to RS_PLUG/RS_UNPLUG masks
The RS_PLUG and RS_UNPLUG interrupt masks are always written as 1 so
those writes are redundant and can be deleted.

This makes it completely clear in the code that only the TS_PLUG and
TS_UNPLUG masks are being changed.

Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>
Link: https://lore.kernel.org/r/20211116163901.45390-3-rf@opensource.cirrus.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit f2dfbaaa5404cadf70213146a5b4b89b647d9092)

BUG=b:218700194
TEST=Test Audio use cases

Signed-off-by: SHANKAR BOYANE <shankarx.boyane@intel.com>
Change-Id: I2900de8a5e4f14dc9dde257710b8e7d427b2416e