UPSTREAM: drm/omap: Add ovl checking funcs to dispc_ops
In order to be able to dynamically assign overlays to planes we need to
be able to asses the overlay capabilities.

Add a helper function to be able to retrieve the supported capabilities
of an overlay.

And export the function to check if a fourcc is supported on a given
overlay.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>
Reviewed-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20211117141928.771082-4-narmstrong@baylibre.com
(cherry picked from commit 0b0f7282f0c81fcbbf300079eaeabd4dcef63c11)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:227338908
TEST=build test

Backported using forklift.py

Change-Id: I7574be0e269f1322973337d4193bdb794670a886