CHROMIUM: iwl7000: mac80211: avoid integer underflow on bad HE caps
If an AP advertises all 8 chains are invalid in HE, the loop
to find the right number will underflow the 'i' variable,
which is a bit odd (but terminates since u16 >> (255*2) == 0).
Make the loop variable signed to avoid that.

Change-Id: I1dee09c744ea29c5dc3ed7096b19757990439501
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: e00941a3360f905a775c0039bbe40187a97f868a