UPSTREAM: ath11k: calculate the correct NSS of peer for HE capabilities
When connected to 6G mode AP, it does not have VHT/HT capabilities,
so the NSS is not set, then it is 1 by default.

This patch is to calculate the NSS with supported HE-MCS and NSS set
of HE capabilities.

Tested-on: WCN6855 hw2.0 PCI WLAN.HSP.1.1-01280-QCAHSPSWPL_V1_V2_SILICONZ_LITE-1

Signed-off-by: Wen Gong <quic_wgong@quicinc.com>
Signed-off-by: Kalle Valo <kvalo@codeaurora.org>
Link: https://lore.kernel.org/r/20211118095453.8030-1-quic_wgong@quicinc.com
(cherry picked from commit 3db26ecf7114370e451e296e33a0af3303d32819)

BUG=b:232096343
TEST=emerge-herobrine chromeos-kernel-5_15; deploy and test basic WiFi
     functionality on herobrine device

Signed-off-by: Kevin Lund <kglund@google.com>
Change-Id: I43568c832bbc626f0658e2b6f4134415572d37c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3654295
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>