CHROMIUM: iwl7000: mvm: use values instead of bits in macros
We get the HE Common Nominal Packet Padding field from the
HE Capabilities IE. Currently we're extracting the bits of this field from
the bitmap but not shifting them, and therefore the macros used to check
the value of this field are bits and not the values of this fields.
In EHT Capabilities IE the position of this field is different,
and this prevent us from using the same macros for both HE and EHT,
which causes to duplicate functions for each one of them.
Fix this by changing the macros to the fields possible values and
by shifting the bits of the field.
In addition, fix typo NOMIMAL - > NOMINAL

Change-Id: I2ca9faf06e177e3035bec91e2ae53c2f91d41774
Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
iwl7000-tree: 733bc886a4268a609635fae51985d3da8db83834