UPSTREAM: net: ipa: zero unused portions of filter table memory
Not all filter table entries are used.  Only certain endpoints
support filtering, and the table begins with a bitmap indicating
which endpoints use the "slots" that follow for filter rules.

Currently, unused filter table entries are not initialized.
Instead, zero-fill the entire unused portion of the filter table
memory regions, to make it more obvious that memory is unused (and
not subsequently modified).

This is not strictly necessary, but the result is reassuring when
looking at filter table memory.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit dc901505fd9820e2428dabb38f5c700d195ce2d3)

BUG=b:200593656
TEST=modem connects on lazor

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: I74a8a064bb54eb41b40c01f9c03295b44ab0cccc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3389837
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>