UPSTREAM: net: ipa: rearrange GSI structure fields
The dummy net_device is a large field in the GSI structure, but it
is not at all interesting from the perspective of debugging.  Move
it to the end of the GSI structure so the other fields are easier to
find in memory.

The channel and event ring arrays are also very large, so move them
near the end of the structure as well.

Swap the position of the result and completion fields to improve
structure packing.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit faa88ecead2f3f1326d6c753af6e700a49872482)

BUG=b:200593656
TEST=modem connects on lazor

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: I6c37bd294df3685a2d3e85a06446bf1b516a4c1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3389842
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>