UPSTREAM: cfg80211: Use the HE operation IE to determine a 6GHz BSS channel
A non-collocated AP whose primary channel is not a PSC channel
may transmit a duplicated beacon on the corresponding PSC channel
in which it would indicate its true primary channel.
Use this inforamtion contained in the HE operation IE to determine
the primary channel of the AP.
In case of invalid infomration ignore it and use the channel
the frame was received on.

Signed-off-by: Ayala Beker <ayala.beker@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Link: https://lore.kernel.org/r/iwlwifi.20211202143322.71eb2176e54e.I130f678e4aa390973ab39d838bbfe7b2d54bff8e@changeid
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit 7f599aeccbd2bcba800c6c7ecc4586fd8cafc1d8)

BUG=b:291984118, b:294242167
TEST=Testing in progress.

Change-Id: I130f678e4aa390973ab39d838bbfe7b2d54bff8e
Signed-off-by: David Ruth <druth@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4749659
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>