UPSTREAM: net: wwan: make debugfs optional
Debugfs interface is optional for the regular modem use. Some distros
and users will want to disable this feature for security or kernel
size reasons. So add a configuration option that allows to completely
disable the debugfs interface of the WWAN devices.

A primary considered use case for this option was embedded firmwares.
For example, in OpenWrt, you can not completely disable debugfs, as a
lot of wireless stuff can only be configured and monitored with the
debugfs knobs. At the same time, reducing the size of a kernel and
modules is an essential task in the world of embedded software.
Disabling the WWAN and IOSM debugfs interfaces allows us to save 50K
(x86-64 build) of space for module storage. Not much, but already
considerable when you only have 16MB of storage.

So it is hard to just disable whole debugfs. Users need some fine
grained set of options to control which debugfs interface is important
and should be available and which is not.

The new configuration symbol is enabled by default and is hidden under
the EXPERT option. So a regular user would not be bothered by another
one configuration question. While an embedded distro maintainer will be
able to a little more reduce the final image size.

(cherry picked from commit 283e6f5a8166f075eba78da6b867d76cc5d47e77)

BUG=b:215304027
TEST=Manual connectivity/traffic tests on Brya and build with W=1

Signed-off-by: Sergey Ryazanov <ryazanov.s.a@gmail.com>
Reviewed-by: Leon Romanovsky <leonro@nvidia.com>
Reviewed-by: Loic Poulain <loic.poulain@linaro.org>
Acked-by: M Chetan Kumar <m.chetan.kumar@intel.com>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
Signed-off-by: Dinesh Kumar Sharma <dinesh.sharma@intel.corp-partner.google.com>
Change-Id: I859f8fdfbc0476c088171847f3d133a75899715c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3933154
Commit-Queue: Daniel Winkler <danielwinkler@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Daniel Winkler <danielwinkler@google.com>
Tested-by: Daniel Winkler <danielwinkler@google.com>
Tested-by: Shrikant Patro <shrikant.patro@intel.corp-partner.google.com>