UPSTREAM: drm/msm/a6xx: Skip crashdumper state if GPU needs_hw_init
I am seeing some crash logs which imply that we are trying to use
crashdumper hw to read back GPU state when the GPU isn't initialized.
This doesn't go well (for example, GPU could be in 32b address mode
and ignoring the upper bits of buffer that it is trying to dump state
to).

I'm not *quite* sure how we get into this state in the first place,
but lets not make a bad situation worse by triggering iova fault
crashes.

While we're at it, also add the information about whether the GPU is
initialized to the devcore dump to make this easier to see in the
logs (which makes the WARN_ON() redundant and even harmful because
it fills up the small bit of dmesg we get with the crash report).

Signed-off-by: Rob Clark <robdclark@chromium.org>
Link: https://lore.kernel.org/r/20211209193118.1163248-1-robdclark@gmail.com
Signed-off-by: Rob Clark <robdclark@chromium.org>
(cherry picked from commit 08c4aa3ee26445ae0d027c1e1944e1202687b024)

BUG=b:222511377
TEST=emerge-trogdor chromeos-kernel-5_15

Signed-off-by: Matt Turner <msturner@google.com>
Change-Id: Ibb0150d910576b1e9ea77a683760a192b43994ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3537861
Reviewed-by: Rob Clark <robdclark@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>