UPSTREAM: phy: qcom-qmp: Add SM8450 USB QMP PHYs
Add support for the USB DP & UNI PHYs found on SM8450. This is same as
the phy version used on SM8350 and sequences turned out to be same, so
use the same table from SM8350 for this as well.

Signed-off-by: Vinod Koul <vkoul@kernel.org>
Reviewed-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Link: https://lore.kernel.org/r/20211213131450.535775-3-vkoul@kernel.org
Signed-off-by: Vinod Koul <vkoul@kernel.org>
(cherry picked from commit 6ad102e05d211aba0ee9c811936eda4341ee5a75)

BUG=b:232267260
TEST=Future patches easier to pick

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: Iacd2c70e6fb4c1cdb76a90ae9bdf65ce01b98026
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3751497
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>