UPSTREAM: thunderbolt: Add CL0s support for USB4 routers
In this patch we add enabling of CL0s - a low power state of the link.
Low power states (called collectively CLx) are used to reduce
transmitter and receiver power when a high-speed lane is idle. For now,
we add support only for first low power state: CL0s.  We enable it, if
both sides of the link support it, and only for the first hop router.
(i.e. the first device that connected to the host router). This is
needed for better thermal management.

Signed-off-by: Gil Fine <gil.fine@intel.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit 8a90e4fa3b4d6a10c96b0a04abf57bb5a16a8f4e)

BUG=b:242112030
TEST=Verify TBT3/USB4 enumeration with USB/DP/PCIe tunnels working.

Change-Id: Iba3b5d59e822aaadbadcb9c80724583f30f1a720
Signed-off-by: Utkarsh Patel <utkarsh.h.patel@intel.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4063212
Reviewed-by: Prashant Malani <pmalani@chromium.org>