UPSTREAM: firmware: cs_dsp: Clear core reset for cache
If the Halo registers are kept in the register cache the
HALO_CORE_RESET bit will be retained as 1 after reset is triggered in
cs_dsp_halo_start_core. This will cause subsequent writes to reset
the core which is not desired. Apart from this bit the rest of the
register bits are cacheable, so for safety sake clear the bit to
ensure the cache is consistent.

Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>
Link: https://lore.kernel.org/r/20220105113026.18955-6-ckeepax@opensource.cirrus.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 7aa1cc1091e0a424e9e7711ca381ebe98b6865bc)

BUG=b:275655477
TEST=power cycle vell amp

Change-Id: I214c5ce8504874c2f8aa7be24df5ed0b52f5630c
Signed-off-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4408188
Commit-Queue: Lalith Rajendran <lalithkraj@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Lalith Rajendran <lalithkraj@google.com>