UPSTREAM: gpu: host1x: Show all allocated syncpts via debugfs
When the host1x syncpts status is dumped via the debugfs, syncpts that
have been allocated but not yet used are not shown and so currently it
is not possible to see all the allocated syncpts. Update the path for
dumping the syncpt status via the debugfs to show all allocated syncpts
even if they have not been used yet. Note that when the syncpt status
is dumped by the kernel itself for debugging only the active syncpt are
shown.

Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>
(cherry picked from commit 74bb98dd91106fcfd16aad852fafa5970a268427)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:242065767
TEST=build test

Backported using forklift.py

Change-Id: Ia9f0e4d1b7d74f548d7be7cf08194283e5196358