CHROMIUM: iwl7000: chromeOS: update and use own ieee80211_data_to_8023_exthdr()
My previous commit adjust checksumming broke on ChromeOS
because one side of the checksum changes was in the driver,
and the other side in cfg80211. However, ChromeOS has their
version of cfg80211, not ours, and so the adjustments were
only partially done.

To fix this, again import ieee80211_data_to_8023_exthdr()
from cfg80211 into our backports so that our mac80211 will
use the correct version of it, the one that assumes the
checksum is including the SNAP header and removes it.

Change-Id: I6055c58f250b4578cb94fdd5d50681346a9c437e
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: e769ed1a940bd6a0c66b9b0250d39230181a60c4