UPSTREAM: drm/vc4: hdmi: Change CSC callback prototype
In order to support the YUV output, we'll need the atomic state to know
what is the state of the associated property in the CSC setup callback.

Let's change the prototype of that callback to allow us to access it.

Acked-by: Thomas Zimmermann <tzimmermann@suse.de>
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Link: https://patchwork.freedesktop.org/patch/msgid/20220120151625.594595-11-maxime@cerno.tech
(cherry picked from commit a34b14a29d96db74876d2cbb466bbc2b73002f3d)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:235993998
TEST=build test

Backported using forklift.py

Change-Id: Ie0e9355d1e3f4e727af9aeb27477d971e5605de2