UPSTREAM: platform/x86: amd-pmc: Correct usage of SMU version
Yellow carp has been outputting versions like `1093.24.0`, but this
is supposed to be 69.24.0. That is the MSB is being interpreted
incorrectly.

The MSB is not part of the major version, but has generally been
treated that way thus far.  It's actually the program, and used to
distinguish between two programs from a similar family but different
codebase.

Link: https://patchwork.freedesktop.org/patch/469993/
Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>
Link: https://lore.kernel.org/r/20220120174439.12770-1-mario.limonciello@amd.com
Reviewed-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>
(cherry picked from commit b8fb0d9b47660ddb8a8256412784aad7cee9f21a)

BUG=b:263448869
TEST=emerge-skyrim-chausie chromeos-kernel-5_15

Change-Id: Ie89001d05857911aa046b5ff887b1648b0c30686
Signed-off-by: Harsh Jain <harsh.jain@amd.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3750785
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Raul Rangel <rrangel@chromium.org>