UPSTREAM: thermal: intel: hfi: Notify user space for HFI events
When the hardware issues an HFI event, relay a notification to user space.
This allows user space to respond by reading performance and efficiency of
each CPU and take appropriate action.

For example, when the performance and efficiency of a CPU is 0, user space
can either offline the CPU or inject idle. Also, if user space notices a
downward trend in performance, it may proactively adjust power limits to
avoid future situations in which performance drops to 0.

To avoid excessive notifications, the rate is limited by one HZ per event.
To limit the netlink message size, send parameters for up to 16 CPUs in a
single message. If there are more than 16 CPUs, issue as many messages as
needed to notify the status of all CPUs.

In the HFI specification, both performance and efficiency capabilities are
defined in the [0, 255] range. The existing implementations of HFI hardware
do not scale the maximum values to 255. Since userspace cares about
capability values that are either 0 or show a downward/upward trend, this
fact does not matter much. Relative changes in capabilities are enough. To
comply with the thermal netlink ABI, scale both performance and efficiency
capabilities to the [0, 1023] interval.

Reviewed-by: Len Brown <len.brown@intel.com>
Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit bd30cdfd9bd73b68e4977ce7c5540aa7b14c25cd)

BUG=b:242115011
TEST=emerge-rex chromeos-kernel-5_15

Signed-off-by: Sachin Kumar <sachin.kumar@intel.com>
Change-Id: Ia9d23bbfa05a68c4b72900ac8e41684c6e9fa736
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3912189
Reviewed-by: Joel Fernandes <joelaf@google.com>
Tested-by: Altaf Basha <altaf.basha@intel.com>
Commit-Queue: Rajat Jain <rajatja@google.com>