UPSTREAM: dt-bindings: usb: dwc3: Deprecate snps,ref-clock-period-ns
This property is redundant because we can determine the correct value for
REFCLKPER based on the "ref" clock. Deprecate it, and encourage users to
provide a clock instead. This also restricts the minimum and maximum to the
values documented in the register reference [1].

[1] https://www.xilinx.com/html_docs/registers/ug1087/usb3_xhci___guctl.html

Reviewed-by: Rob Herring <robh@kernel.org>
Signed-off-by: Sean Anderson <sean.anderson@seco.com>
Link: https://lore.kernel.org/r/20220127200636.1456175-2-sean.anderson@seco.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 97ecbfb30dca36843d7a456def55e4d19007c3d2)

BUG=b:243399729
TEST=USB still works on trogdor and herobrine

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: I942179b0800059838078a56de6d9cfa93fccee05
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3851970
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Douglas Anderson <dianders@chromium.org>