UPSTREAM: usb: dwc3: Calculate REFCLKPER based on reference clock
Instead of using a special property to determine the reference clock
period, use the rate of the reference clock. When we have a legacy
snps,ref-clock-period-ns property and no reference clock, use it
instead. Fractional clocks are not currently supported, and will be
dealt with in the next commit.

Tested-by: Robert Hancock <robert.hancock@calian.com>
Reviewed-by: Robert Hancock <robert.hancock@calian.com>
Reviewed-by: Thinh Nguyen <Thinh.Nguyen@synopsys.com>
Signed-off-by: Sean Anderson <sean.anderson@seco.com>
Link: https://lore.kernel.org/r/20220127200636.1456175-4-sean.anderson@seco.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 5114c3ee24875270e9f9ef696c46832e5782587c)

BUG=b:243399729
TEST=USB still works on trogdor and herobrine

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: Icd324ddef6e8c9b4d27e17f41adff2d52058e0d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3851972
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>