UPSTREAM: drm/exynos: fimd: add BGR support for exynos4/5
In the downstream kernels for exynos4 and exynos5 devices, there is an
undocumented register that controls the order of the RGB output. It can
be set to either normal order or reversed, which enables BGR support for
those SoCs.

This patch enables the BGR support for all the SoCs that were found to
have at least one device with this logic in the corresponding downstream
kernels.

Signed-off-by: Martin Jücker <martin.juecker@gmail.com>
Signed-off-by: Inki Dae <inki.dae@samsung.com>
(cherry picked from commit 2d684f4e155c1e80ff63bd503930171c460eac5b)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:235993998
TEST=build test

Backported using forklift.py

Change-Id: Ib2f8979d07327a8b9746136b0019ad6e487c5e34