UPSTREAM: dt-bindings: net: qcom,ipa: add optional qcom,qmp property
For some systems, the IPA driver must make a request to ensure that
its registers are retained across power collapse of the IPA hardware.
On such systems, we'll use the existence of the "qcom,qmp" property
as a signal that this request is required.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit ac62a0174d62ae0f4447c0c8cf35a8e5d793df56)

BUG=b:200593656
TEST=emerge-herobrine sys-kernel/chromeos-kernel-5_15

Change-Id: Ie96b051b3347b0792d9d628e34313846636be180
Signed-off-by: Atul Dhudase <adhudase@codeaurora.org>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3481158
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>