UPSTREAM: net: ipa: kill replenish_saved
The replenish_saved field keeps track of the number of times a new
buffer is added to the backlog when replenishing is disabled.  We
don't really use it though, so there's no need for us to track it
separately.  Whether replenishing is enabled or not, we can simply
increment the backlog.

Get rid of replenish_saved, and initialize and increment the backlog
where it would have otherwise been used.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit a9bec7ae70c16f003d6138617c8bd305de3b230e)

BUG=b:228884730
TEST=Tested on Lazor. Wifi, LTE, and bluetooth all still work.

Signed-off-by: Joseph S. Barrera III <joebar@chromium.org>
Change-Id: I704aef6b5c26df8f49d0a4d7c400030986829129
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3767080
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>