UPSTREAM: net: ipa: allocate transaction in replenish loop
When replenishing, have ipa_endpoint_replenish() allocate a
transaction, and pass that to ipa_endpoint_replenish_one() to fill.
Then, if that produces no error, commit the transaction within the
replenish loop as well.  In this way we can distinguish between
transaction failures and buffer allocation/mapping failures.

Failure to allocate a transaction simply means the hardware already
has as many receive buffers as it can hold.  In that case we can
break out of the replenish loop because there's nothing more to do.

If we fail to allocate or map pages for the receive buffer, just
try again later.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 6a606b90153b821915daade0b8f253d01d443d75)

BUG=b:228884730
TEST=Tested on Lazor. Wifi, LTE, and bluetooth all still work.

Signed-off-by: Joseph S. Barrera III <joebar@chromium.org>
Change-Id: Ic0e42accdbee39c633dce87ab812df2e73fb0627
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3767084
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>