FROMGIT: drm/panel-edp: Allow querying the detected panel via debugfs
Recently we added generic "edp-panel"s probed by EDID. To support
panels in this way we look at the panel ID in the EDID and look up the
panel in a table that has power sequence timings. If we find a panel
that's not in the table we will still attempt to use it but we'll use
conservative timings. While it's likely that these conservative
timings will work for most nearly all panels, the performance of
turning the panel off and on suffers.

We'd like to be able to reliably detect the case that we're using the
hardcoded timings without relying on parsing dmesg. This allows us to
implement tests that ensure that no devices get shipped that are
relying on the conservative timings.

Let's add a new debugfs entry to panel devices. It will have one of:
* UNKNOWN - We tried to detect a panel but it wasn't in our table.
* HARDCODED - We're not using generic "edp-panel" probed by EDID.
* A panel name - This is the name of the panel from our table.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Javier Martinez Canillas <javierm@redhat.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20220204161245.v2.3.I209d72bcc571e1d7d6b793db71bf15c9c0fc9292@changeid
(cherry picked from commit 6ed19359d6bd62e993b09a7a565d7a5ce5e114c3
 git://anongit.freedesktop.org/drm/drm-misc drm-misc-next)

BUG=b:215432805
TEST=grep "" /sys/kernel/debug/dri/*/eDP*/panel/detected_panel

Change-Id: I209d72bcc571e1d7d6b793db71bf15c9c0fc9292
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3470722
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Tested-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>