UPSTREAM: fbdev/defio: Early-out if page is already enlisted
Return early if a page is already in the list of dirty pages for
deferred I/O. This can be detected if the page's list head is not
empty. Keep the list head initialized while the page is not enlisted
to make this work reliably.

v2:
	* update comment and fix spelling (Sam)

Signed-off-by: Thomas Zimmermann <tzimmermann@suse.de>
Acked-by: Sam Ravnborg <sam@ravnborg.org>
Link: https://patchwork.freedesktop.org/patch/msgid/20220211094640.21632-2-tzimmermann@suse.de
(cherry picked from commit 105a940416fc622406653b6fe54732897642dfbc)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:235993998
TEST=build test

Backported using forklift.py

Change-Id: Id6888c406cffbdd3e1e253423103c155ff3cf88c