FIXUP: FROMLIST: HID: haptic: initialize haptic device
There was an 'off-by-one' bug in hid_haptic_upload_effect() which caused
that the effect for the last waveform could not be uploaded.
When iterating over ordinals let the last one be equal max_waveform_id.

Signed-off-by: Sean O'Brien <seobrien@chromium.org>

BUG=b:208873623
TEST=Build, update kernel and make sure Redrix boots fine. Try to upload
an effect for the last waveform.
UPSTREAM-TASK=b:287965840

Change-Id: I5e6c143d83640cbafd75ba697ed6b233e59f07e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/4627037
Commit-Queue: Sean O'Brien <seobrien@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Tested-by: Sean O'Brien <seobrien@chromium.org>