CHROMIUM: iwl7000: mvm: add an option to set the minimum allowed beacon interval
The driver does not associate to APs with a beacon interval shorter
than 16 TUs because of an old fw bug. Make this configurable since
it may be needed to connect to APs with shorter beacon interval and
the fw bug does not exist anymore.

Change-Id: Ie37ccda2aa51428f76a91f4ef8173a9b361bea55
Signed-off-by: Avraham Stern <avraham.stern@intel.com>
iwl7000-tree: 2fd8d83664fc0c8969943758409ba55e28f5f7b3