UPSTREAM: drm/vc4: hdmi: Always try to have the highest bpc
Currently we take the max_bpc property as the bpc value and do not try
anything else.

However, what the other drivers seem to be doing is that they would try
with the highest bpc allowed by the max_bpc property and the hardware
capabilities, test if it results in an acceptable configuration, and if
not decrease the bpc and try again.

Let's use the same logic.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Acked-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20220222164042.403112-7-maxime@cerno.tech
(cherry picked from commit dd39d024a8471e88951f8494f19e693c9e3185a9)
Signed-off-by: Drew Davenport <ddavenport@chromium.org>

BUG=b:242065767
TEST=build test

Backported using forklift.py

Change-Id: Ia5003c7398a980f99ed0e1e97cd6dfca57034480