FROMGIT: drm/msm: Add MSM_SUBMIT_FENCE_SN_IN
Add a way for userspace to specify the sequence number fence used to
track completion of the submit.  As the seqno fence is simply an
incrementing counter which is local to the submitqueue, it is easy for
userspace to know the next value.

This is useful for native userspace drivers in a vm guest, as the guest
to host roundtrip can have high latency.  Assigning the fence seqno in
the guest userspace allows the guest to continue without waiting for
response from the host.

Signed-off-by: Rob Clark <robdclark@chromium.org>
Link: https://lore.kernel.org/r/20220224222321.60653-1-robdclark@gmail.com
(cherry picked from commit 17154addc5c1a175bcf3441ff0d9598efa1f05cd
 https://gitlab.freedesktop.org/drm/msm.git msm-next)

BUG=b:222511377
TEST=emerge-trogdor chromeos-kernel-5_15

Signed-off-by: Matt Turner <msturner@google.com>
Change-Id: I7f2b2c19211c0de988f2abe7ce09b7053498f2ab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3541294
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Rob Clark <robdclark@chromium.org>