UPSTREAM: thunderbolt: Do not make DROM read success compulsory
The USB4 specification doesn't make any requirements that reading
a device router's DROM is needed for the operation of the device.

Other connection manager solutions don't necessarily read it or gate
the usability of the device on whether it was read.

So make failures when reading the DROM show warnings but not
fail the initialization of the router.

Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit 6915812bbd109787ebdb865561dc9164d4b01f56
 https://git.kernel.org/pub/scm/linux/kernel/git/westeri/thunderbolt.git next)

BUG=b:242112030
TEST=emerge-rex sys-kernel/chromeos-kernel-5_15

Signed-off-by: Narasimha Rao Rayala <narasimha.rao.rayala@intel.com>
Change-Id: Ic08cfb19f66c982cd4fdda62b424ec2cb5d3a255
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3887001
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>